LocalTest/requirements.txt
aiohttp==3.9.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.6
distlib==0.3.8
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.4
frozenlist==1.4.1
fsspec==2023.4.0
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-spreadsheet==0.0.6
googleapis-common-protos==1.63.0
grpcio==1.62.1
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.7
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
packaging==24.0
pandas==2.2.2
pillow==10.2.0
platformdirs==4.2.0
prettytable==3.10.0
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
py-spy==0.3.14
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ray==2.10.0
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tensorboardX==2.6.2.2
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
